...Well... It works I guess?

+++ Utils.cpp
@@ -1,14 +1,14 @@
 #include "Utils.h"
 
-bool Utils::pointOnObject(sf::Vector2i pos, GameObject* obj)
+bool Utils::pointOnObject(sf::Vector2f pos, GameObject* obj)
 {
     int p_x = pos.x;
     int p_y = pos.y;
 
-    sf::Vector2i vect_pos = obj->getPos();
+    sf::Vector2f vect_pos = obj->getPos();
     int o_x = vect_pos.x;
     int o_y = vect_pos.y;
-    sf::Vector2i vect_size = obj->getSize();
+    sf::Vector2f vect_size = obj->getSize();
     int o_w = vect_size.x;
     int o_h = vect_size.y;
